+++ plot/image.go
@@ -7,36 +7,22 @@ import (
 	"github.com/faiface/pixel/pixelgl"
 	"github.com/mazznoer/colorgrad"
 	"github.com/mlange-42/arche-model/observer"
-	"github.com/mlange-42/arche-pixel/window"
 	"github.com/mlange-42/arche/ecs"
 )
 
-// Image plot reporter.
-//
-// If the world contains a resource of type [github.com/mlange-42/arche-model/resource.Termination],
-// the model is terminated when the window is closed.
+// Image plot drawer.
 type Image struct {
-	Bounds         window.Bounds      // Window bounds.
-	Scale          float64            // Spatial scaling: cell size in screen pixels.
-	Observer       observer.Matrix    // The observer.
-	Colors         colorgrad.Gradient // Colors for mapping values.
-	Min            float64            // Minimum value for color mapping.
-	Max            float64            // Maximum value for color mapping. Is set to 1.0 if both Min and Max are zero.
-	UpdateInterval int                // Interval for updating the observer, in model ticks.
-	DrawInterval   int                // Interval for re-drawing, in UI frames.
-	Drawers        []window.Drawer    // Optional additional drawers for drawing over the image.
-	window         window.Window
-	drawer         imageDrawer
-	step           int64
+	Scale    float64            // Spatial scaling: cell size in screen pixels.
+	Observer observer.Matrix    // Observer providing 2D matrix or grid data.
+	Colors   colorgrad.Gradient // Colors for mapping values.
+	Min      float64            // Minimum value for color mapping.
+	Max      float64            // Maximum value for color mapping. Is set to 1.0 if both Min and Max are zero.
+	slope    float64
+	picture  *pixel.PictureData
 }
 
-// Initialize the system.
-func (s *Image) Initialize(w *ecs.World) {
-	s.step = 0
-}
-
-// InitializeUI the system.
-func (s *Image) InitializeUI(w *ecs.World) {
+// Initialize the system
+func (s *Image) Initialize(w *ecs.World, win *pixelgl.Window) {
 	s.Observer.Initialize(w)
 
 	if s.Scale <= 0 {
@@ -46,68 +32,20 @@ func (s *Image) InitializeUI(w *ecs.World) {
 		s.Max = 1
 	}
 
-	s.drawer = newImageDrawer(s.Observer, s.Scale, &s.Colors, s.Min, s.Max)
+	s.slope = 1.0 / (s.Max - s.Min)
 
-	s.window.DrawInterval = s.DrawInterval
-	s.window.Bounds = s.Bounds
-	s.window.Drawers = append([]window.Drawer{&s.drawer}, s.Drawers...)
-	s.window.InitializeUI(w)
-}
-
-// Update the system.
-func (s *Image) Update(w *ecs.World) {
-	if s.UpdateInterval <= 1 || s.step%int64(s.UpdateInterval) == 0 {
-		s.Observer.Update(w)
-	}
-	s.step++
-}
-
-// UpdateUI the system.
-func (s *Image) UpdateUI(w *ecs.World) {
-	s.window.UpdateUI(w)
-}
-
-// PostUpdateUI updates the GL window.
-func (s *Image) PostUpdateUI(w *ecs.World) {
-	s.window.PostUpdateUI(w)
-}
-
-// Finalize the system.
-func (s *Image) Finalize(w *ecs.World) {}
-
-// FinalizeUI the system.
-func (s *Image) FinalizeUI(w *ecs.World) {
-	s.window.FinalizeUI(w)
-}
-
-type imageDrawer struct {
-	observer observer.Matrix
-	scale    float64
-	colors   *colorgrad.Gradient
-	offset   float64
-	slope    float64
-	picture  *pixel.PictureData
-}
-
-func newImageDrawer(obs observer.Matrix, scale float64, colors *colorgrad.Gradient, min, max float64) imageDrawer {
-	return imageDrawer{
-		observer: obs,
-		scale:    scale,
-		colors:   colors,
-		offset:   min,
-		slope:    1.0 / (max - min),
-	}
-}
-
-// Initialize the system
-func (s *imageDrawer) Initialize(w *ecs.World, win *pixelgl.Window) {
-	width, height := s.observer.Dims()
+	width, height := s.Observer.Dims()
 	s.picture = pixel.MakePictureData(pixel.R(0, 0, float64(width), float64(height)))
 }
 
+// Update the drawer.
+func (s *Image) Update(w *ecs.World) {
+	s.Observer.Update(w)
+}
+
 // Draw the system
-func (s *imageDrawer) Draw(w *ecs.World, win *pixelgl.Window) {
-	values := s.observer.Values(w)
+func (s *Image) Draw(w *ecs.World, win *pixelgl.Window) {
+	values := s.Observer.Values(w)
 
 	length := len(values)
 	for i := 0; i < length; i++ {
@@ -115,11 +53,11 @@ func (s *imageDrawer) Draw(w *ecs.World, win *pixelgl.Window) {
 	}
 
 	sprite := pixel.NewSprite(s.picture, s.picture.Bounds())
-	sprite.Draw(win, pixel.IM.Moved(pixel.V(s.picture.Rect.W()/2.0, s.picture.Rect.H()/2.0)).Scaled(pixel.Vec{}, s.scale))
+	sprite.Draw(win, pixel.IM.Moved(pixel.V(s.picture.Rect.W()/2.0, s.picture.Rect.H()/2.0)).Scaled(pixel.Vec{}, s.Scale))
 }
 
-func (s *imageDrawer) valueToColor(v float64) color.RGBA {
-	c := s.colors.At((v - s.offset) * s.slope)
+func (s *Image) valueToColor(v float64) color.RGBA {
+	c := s.Colors.At((v - s.Min) * s.slope)
 	return color.RGBA{
 		R: uint8(c.R * 255),
 		G: uint8(c.G * 255),