+++ plot/image_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/mlange-42/arche-model/model"
 	"github.com/mlange-42/arche-model/system"
 	"github.com/mlange-42/arche-pixel/plot"
+	"github.com/mlange-42/arche-pixel/window"
 	"github.com/mlange-42/arche/ecs"
 )
 
@@ -21,15 +22,17 @@ func ExampleImage() {
 
 	// Create an image plot.
 	// See below for the implementation of the MatrixObserver.
-	pl := plot.Image{
-		Scale:    4,
-		Observer: &MatrixObserver{},
-		Colors:   colorgrad.Inferno(),
-		Min:      -2,
-		Max:      2,
-	}
-	// Add the plot as UI system.
-	m.AddUISystem(&pl)
+	m.AddUISystem(&window.Window{
+		Drawers: []window.Drawer{
+			&plot.Image{
+				Scale:    4,
+				Observer: &MatrixObserver{},
+				Colors:   colorgrad.Inferno(),
+				Min:      -2,
+				Max:      2,
+			},
+		},
+	})
 
 	// Add a termination system that ends the simulation.
 	m.AddSystem(&system.FixedTermination{
@@ -40,7 +43,7 @@ func ExampleImage() {
 	// Due to the use of the OpenGL UI system, the model must be run via [github.com/faiface/pixel/pixelgl].
 	// Uncomment the next line. It is commented out as the CI has no display device to test the model run.
 
-	//pixelgl.Run(m.Run)
+	// pixelgl.Run(m.Run)
 }
 
 // Example observer, reporting a matrix with z = sin(0.1*i) + sin(0.2*j).