+++ plot/time_series_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/mlange-42/arche-model/model"
 	"github.com/mlange-42/arche-model/system"
 	"github.com/mlange-42/arche-pixel/plot"
+	"github.com/mlange-42/arche-pixel/window"
 	"github.com/mlange-42/arche/ecs"
 )
 
@@ -20,11 +21,13 @@ func ExampleTimeSeries() {
 
 	// Create a time series plot.
 	// See below for the implementation of the RowObserver.
-	pl := plot.TimeSeries{
-		Observer: &RowObserver{},
-	}
-	// Add the plot as UI system.
-	m.AddUISystem(&pl)
+	m.AddUISystem(&window.Window{
+		Drawers: []window.Drawer{
+			&plot.TimeSeries{
+				Observer: &RowObserver{},
+			},
+		},
+	})
 
 	// Add a termination system that ends the simulation.
 	m.AddSystem(&system.FixedTermination{
@@ -35,7 +38,7 @@ func ExampleTimeSeries() {
 	// Due to the use of the OpenGL UI system, the model must be run via [github.com/faiface/pixel/pixelgl].
 	// Uncomment the next line. It is commented out as the CI has no display device to test the model run.
 
-	//pixelgl.Run(m.Run)
+	// pixelgl.Run(m.Run)
 }
 
 // RowObserver to generate random time series.