+++ window/example_test.go
@@ -51,6 +51,9 @@ func (d *RectDrawer) Initialize(w *ecs.World, win *pixelgl.Window) {
 	d.dr = *imdraw.New(nil)
 }
 
+// Update the RectDrawer (does nothing).
+func (d *RectDrawer) Update(w *ecs.World) {}
+
 // Draw the RectDrawer's stuff.
 func (d *RectDrawer) Draw(w *ecs.World, win *pixelgl.Window) {
 	// Get a resource from the world.