+++ window/window.go
@@ -30,6 +30,7 @@ func B(x, y, w, h int) Bounds {
 // See the example for [window] package.
 type Drawer interface {
 	Initialize(w *ecs.World, win *pixelgl.Window)
+	Update(w *ecs.World)
 	Draw(w *ecs.World, win *pixelgl.Window)
 }
 
@@ -40,18 +41,25 @@ type Drawer interface {
 //
 // See the example for [window] package.
 type Window struct {
-	Bounds       Bounds
-	DrawInterval int
-	Drawers      []Drawer
-	window       *pixelgl.Window
-	step         int
-	isClosed     bool
-	termRes      generic.Resource[resource.Termination]
+	Bounds         Bounds
+	Drawers        []Drawer
+	UpdateInterval int // Interval for updating the observer, in model ticks.
+	DrawInterval   int // Interval for re-drawing, in UI frames.
+	window         *pixelgl.Window
+	updateStep     int64
+	drawStep       int64
+	isClosed       bool
+	termRes        generic.Resource[resource.Termination]
 }
 
 // AddDrawer adds a drawer.
 func (s *Window) AddDrawer(d Drawer) {
 	s.Drawers = append(s.Drawers, d)
+}
+
+// Initialize the system.
+func (s *Window) Initialize(w *ecs.World) {
+	s.updateStep = 0
 }
 
 // InitializeUI the system.
@@ -88,8 +96,21 @@ func (s *Window) InitializeUI(w *ecs.World) {
 	}
 
 	s.termRes = generic.NewResource[resource.Termination](w)
-	s.step = 0
+	s.drawStep = 0
 	s.isClosed = false
+}
+
+// Update the system.
+func (s *Window) Update(w *ecs.World) {
+	if s.isClosed {
+		return
+	}
+	if s.UpdateInterval <= 1 || s.updateStep%int64(s.UpdateInterval) == 0 {
+		for _, d := range s.Drawers {
+			d.Update(w)
+		}
+	}
+	s.updateStep++
 }
 
 // UpdateUI the system.
@@ -104,20 +125,23 @@ func (s *Window) UpdateUI(w *ecs.World) {
 		}
 		return
 	}
-	if s.DrawInterval <= 1 || s.step%s.DrawInterval == 0 {
+	if s.DrawInterval <= 1 || s.drawStep%int64(s.DrawInterval) == 0 {
 		s.window.Clear(colornames.Black)
 
 		for _, d := range s.Drawers {
 			d.Draw(w, s.window)
 		}
 	}
-	s.step++
+	s.drawStep++
 }
 
 // PostUpdateUI updates the GL window.
 func (s *Window) PostUpdateUI(w *ecs.World) {
 	s.window.Update()
 }
+
+// Finalize the system.
+func (s *Window) Finalize(w *ecs.World) {}
 
 // FinalizeUI the system.
 func (s *Window) FinalizeUI(w *ecs.World) {}