+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 ### Breaking changes
 
 * Renamed `Window.Add` to `Window.AddDrawer` (#8)
+* `Drawer` interface has method `Update(w *ecs.World)` (#8)
+* All plots are `Drawer` instead of `UISystem`, and are added to a `Window` (#8)
 
 ### Features
 
 * Adds `Image` plot for plotting grids and matrices (#8)
 * Adds `ImageRGB` plot for plotting multi-channel grids and matrices (#8)
-* Additional drawers can be added to `Image` and `ImageRGB` plots, and draw over the plot's drawing (#8)
 
 ## [[v0.0.2]](https://github.com/mlange-42/arche-pixel/compare/v0.0.1...v0.0.2)